backend/requirements.txt
# Core dependencies
fastapi[standard]==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.1

# Database
boto3==1.29.0

# Authentication
python-jose==3.3.0

# WebSocket support
websockets==14.2

# Type support and validation
pydantic==2.10.6

svix==1.16.0 

clerk-backend-api==1.7.2